+++ ScenarioGUI/gui_classes/gui_structure.py
@@ -58,6 +58,11 @@ class GuiStructure:
         FunctionButton.default_parent = default_parent
         self.translations = translations
 
+        # flag for loading the gui
+        self.started: bool = False
+        # flag for loading a file
+        self.loaded: bool = False
+
         # create page
         # self.page_result = None